First commit.
This is a rerendering of the laptop script into Ansible.

There is more that the script did that this does *not* do, but it
transforms the majority of the bash scripting into Ansible playbooks.

The bash script at the start does a bit of shell config, installs
Homebrew, and sets up Python using Homebrew. From there, we fall into
Ansible.

The README and LICENSE will need to be updated. This is just to get the
commit off the local machine.