Pull repos, run playbook.
The original bootstrap did too little.

This version pulls the support scripts (they could all be one file),
sources them, installs git, and then goes.

I'm going to open a ticket, as I'm not confident I should do the Ansible
install globally. I think I'd like to do it in a venv, and be able to
have a clean/temporary touch on the host when I'm done.