Adding tasks for GPG, seekret
I forgot these in the last commit.

The GPG task installs GPG and does some mangling to make it work on
macOS. This is not idempotent, I suspect.

seekret is not yet working, but it uses github3 to pull the most recent
release via API. This is much cleaner than mangling the source of the
webpage.