Moving to the brew version of Ansible.
This eliminates the need to download and install pip. I read some
reports online that the brew version of ansible sometimes can misbehave,
but they were vague statements. It seems to run fine. (Hm.)

Point being, this eliminates an entire set of setup dependencies, and
means that future runs can count on ansible existing. Therefore, a small
wrapper script to 1) checkout the playbook and 2) run ansible can be
dropped somewhere. And, actually, that can be a playbook itself.

So, this improves the repeatability. Perhaps the repos should be checked
out to somewhere like .laptop, and that can be used as a working
directory. To be discussed.