Adding caulking.
resolves #8

Reference: https://github.com/18F/laptop/issues/185

The cloud.gov team uses caulking to manage secrets checking, whereas the
install of git-seekrets I was using was an aging (unmaintained?) fork.
It appears from prior conversation that there was a decision made to use
the tooling used/developed by (?) the cloud.gov team.

The caulking role pulls the repo and runs "make && make install", and
then makes sure patterns are up-to-date. While this has a certain lack
of idempotency, it will make sure that, with every run of the playbook,
we are up-to-date.